Reduce size of generated tuple type definitions (#2283)
Since this PR contains refactorings, please review using the commits. The refactoring commits are all at the start so it should be easy to skip just those and review the remaining changes as one big diff.

### Changes
- Refactoring: move TupleTypeDecl from DafnyAST.cs to its own file
- Change the CompileName TupleTypeDecl so it ignores any ghost arguments of the tuple
- Change SinglePassCompiler so datatypes with the same CompileName only generate one target language type.
- Change `BuiltIns` so tuples with exactly two non-ghost arguments are not compiled

### Testing
This PR only affects the size of the generated code, which we currently do not test, so there are no test changes. The goal of this PR is to prepare for another PR where we no longer generate tuples but instead only support them to some fixes size (say 10) by putting them in the runtime. With this PR, we only need 10 predefined tuples in the runtime, instead of 2^10, and we support more of them (all tuples with 10 non-ghost arguments, instead of all tuples with 10 arguments in total)
